requirements.txt
discord.py[voice]
discord-ext-voice-recv
python-dotenv
websockets
SpeechRecognition